+++ src/app/core/+store/app.state.ts
@@ -1,7 +1,5 @@
 import { ProductsState } from './products';
-import { CartState } from './cart';
 
 export interface AppState {
   products: ProductsState;
-  cart: CartState;
 }