+++ src/app/core/+store/core-store.module.ts
@@ -4,15 +4,26 @@ import { CommonModule } from '@angular/common';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { StoreRouterConnectingModule } from '@ngrx/router-store';
+
 import { environment } from './../../../environments/environment';
+import {
+  RouterStateSerializerProvider,
+  routerReducers,
+  RouterEffects } from './router';
+
 
 @NgModule({
   imports: [
     CommonModule,
-    StoreModule.forRoot({}),
-    EffectsModule.forRoot([]),
+    StoreModule.forRoot(routerReducers),
+    EffectsModule.forRoot([RouterEffects]),
+    StoreRouterConnectingModule.forRoot(),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
   ],
-  declarations: []
+  declarations: [],
+  providers: [
+    RouterStateSerializerProvider
+  ]
 })
 export class CoreStoreModule { }