+++ src/app/core/+store/products/products.actions.ts
@@ -2,12 +2,11 @@ import { Action } from '@ngrx/store';
 
 import { Product } from './../../../products/models/product.model';
 
+
 export enum ProductsActionTypes {
   GET_PRODUCTS = '[Products] GET_PRODUCTS',
   GET_PRODUCTS_SUCCESS = '[Products] GET_PRODUCTS_SUCCESS',
   GET_PRODUCTS_ERROR = '[Products] GET_PRODUCTS_ERROR',
-
-  GET_PRODUCT = '[Products] GET_PRODUCT',
 
   DELETE_PRODUCT = '[Products] DELETE_PRODUCT',
   DELETE_PRODUCT_SUCCESS = '[Products] DELETE_PRODUCT_SUCCESS',
@@ -35,13 +34,6 @@ export class GetProductsSuccess implements Action {
 export class GetProductsError implements Action {
   readonly type = ProductsActionTypes.GET_PRODUCTS_ERROR;
   constructor(public payload: Error | string) { }
-}
-
-
-// GetProduct
-export class GetProduct implements Action {
-  readonly type = ProductsActionTypes.GET_PRODUCT;
-  constructor(public payload: number) { }
 }
 
 
@@ -100,7 +92,6 @@ export type ProductsActions
   = GetProducts
   | GetProductsSuccess
   | GetProductsError
-  | GetProduct
   | DeleteProduct
   | DeleteProductSuccess
   | DeleteProductError