+++ src/app/core/+store/router/router.effects.ts
@@ -1,0 +1,39 @@
+import { Injectable } from '@angular/core';
+import { Router } from '@angular/router';
+import { Location } from '@angular/common';
+
+import { Effect, Actions, ofType } from '@ngrx/effects';
+import { tap, pluck } from 'rxjs/operators';
+
+import * as RouterActions from './router.actions';
+
+
+@Injectable()
+export class RouterEffects {
+  constructor(
+    private actions$: Actions,
+    private router: Router,
+    private location: Location
+  ) {}
+
+  @Effect({ dispatch: false })
+  navigate$ = this.actions$.pipe(
+    ofType<RouterActions.Go>(RouterActions.RouterActionTypes.GO),
+    pluck('payload'),
+    tap(({ path, queryParams, extras }) => {
+      this.router.navigate(path, { queryParams, ...extras });
+    })
+  );
+
+  @Effect({ dispatch: false })
+  navigateBack$ = this.actions$.pipe(
+    ofType<RouterActions.Back>(RouterActions.RouterActionTypes.BACK),
+    tap(() => this.location.back())
+  );
+
+  @Effect({ dispatch: false })
+  navigateForward$ = this.actions$.pipe(
+    ofType<RouterActions.Forward>(RouterActions.RouterActionTypes.FORWARD),
+    tap(() => this.location.forward())
+  );
+}