+++ src/app/core/guards/auth.guard.ts
@@ -4,22 +4,26 @@ import {
   CanLoad,
   ActivatedRouteSnapshot,
   RouterStateSnapshot,
-  Router,
   Route,
   NavigationExtras } from '@angular/router';
+
+import { Store } from '@ngrx/store';
+import { AppState } from './../+store';
+import * as RouterActions from './../+store/router/router.actions';
 
 import { Observable } from 'rxjs';
 
 import { CoreModule } from '../core.module';
 import { AuthService } from './../services';
 
+
 @Injectable({
-  providedIn: 'root'//CoreModule
+  providedIn: 'root'
 })
 export class AuthGuard implements CanActivate, CanLoad {
   constructor(
     private authService: AuthService,
-    private router: Router
+    private store: Store<AppState>
   ) {}
 
   private checkLogin(url: string): boolean {
@@ -29,7 +33,10 @@ export class AuthGuard implements CanActivate, CanLoad {
 
     this.authService.redirectUrl = url;
 
-    this.router.navigate(['/login']);
+    this.store.dispatch(new RouterActions.Go({
+      path: ['/login']
+    }));
+
     return false;
   }